Refactor installation script and update dependencies
This commit refactors the installation script and updates the dependencies. The changes include:
- Installing repository packages using paru
- Installing AUR packages using paru
- Setting the Plymouth theme
- Setting up mkinitcpio to add Plymouth
- Enabling splash in kernel parameters
- Changing the timer for systemd-boot
- Slowing down boot to display Plymouth
- Activating services

Additionally, the commit updates the Pacman configuration, installs needed packages for setup, adds the user to sudoers without password, disables debug for makepkg, installs the AUR helper paru, updates the mirrorlist, sets the default shell to zsh, and sets the SDDM theme.